Removes repetitive code.
