+++ app/services/self_deposit/custom_queries/find_publication_by_deduplication_key.rb
@@ -7,17 +7,10 @@ module SelfDeposit
     #   Hyrax.custom_queries.find_publication_by_deduplication_key(deduplication_key:)
     #
     # @see https://github.com/samvera/valkyrie/wiki/Queries#custom-queries
-    class FindPublicationByDeduplicationKey
+    class FindPublicationByDeduplicationKey < SolrDocumentQuery
       def self.queries
         [:find_publication_by_deduplication_key]
       end
-
-      def initialize(query_service:)
-        @query_service = query_service
-        @connection = Hyrax.index_adapter.connection
-      end
-
-      attr_reader :query_service
 
       ##
       # @param deduplication_key for a Publication
@@ -27,12 +20,6 @@ module SelfDeposit
         @deduplication_key = deduplication_key
         raise ::Valkyrie::Persistence::ObjectNotFoundError unless resource
         @query_service.find_by(id: resource['id'])
-      end
-
-      # Queries Solr for a document that matches the provided key
-      # @yield [Publication]
-      def resource
-        @connection.get("select", params: { q: query, fl: "*", rows: 1 })["response"]["docs"].first
       end
 
       # Solr query for for a Publication with a deduplication_key_tesi that matches the provided key
